move all assertion-related logic from CapturingAppender to LogAsserter
